[STYLE] 파일명 변경
 js -> jsx
